Updated the tab handling to make adding new tabs easier.
Added a handler script to hold the tab data.
Tabline now redirects the call to change activity to the handler script.
Tabs buttons are now their own prefab.
