+++ Assets/MenuUi/Prefabs/UI Components/TabObject.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3a035a942e5180d45947bdb51427e49d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
