+++ Assets/MenuUi/Scripts/TabLine/TabObjectHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 48bd9b2820919a14b9d5fc0f9e5bfe9d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
